Extract renderer from the icon example to make it reusable
